Added some xmldocs
